Put IPU pipelines into separate files
